requirements.txt
streamlit>=1.43.0
groq>=0.20.0
python-dotenv>=1.0.1
vaderSentiment>=3.3.2
